+++ marimo/_plugins/ai/__init__.py
@@ -4,7 +4,7 @@
 from __future__ import annotations
 
 __all__ = [
-    "models",
+    "llm",
 ]
 
-from marimo._plugins.ui._impl.chat import models
+from marimo._plugins.ui._impl.chat import llm
